Fix Scan Line bugs

+++ RawLine/Tools/Aresta.cs
@@ -8,15 +8,15 @@ namespace RawLine.Tools
 {
     class Aresta
     {
-        private int maxY;
-        private int minX;
+        private double maxY;
+        private double minX;
         private double incX;
 
-        public int MaxY { get => this.maxY; set => this.maxY = value; }
-        public int MinX { get => this.minX; set => this.minX = value; }
+        public double MaxY { get => this.maxY; set => this.maxY = value; }
+        public double MinX { get => this.minX; set => this.minX = value; }
         public double IncX { get => this.incX; set => this.incX = value; }
 
-        public Aresta(int maxY, int minX, double incX)
+        public Aresta(double maxY, double minX, double incX)
         {
             this.maxY = maxY;
             this.minX = minX;
@@ -25,7 +25,7 @@ namespace RawLine.Tools
 
         public void Add()
         {
-            this.minX = Convert.ToInt32(this.minX + this.incX);
+            this.minX = Convert.ToDouble(this.minX + this.incX);
         }
     }
 }
